requirements.txt
boto3==1.35.40
botocore==1.35.40
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
urllib3==2.2.3
